+++ CHANGELOG.md
@@ -13,6 +13,9 @@ BikeControl Pro is now available - subscribe to the Pro version to support the p
   - perform steering using long presses
   - gear changes are now reflected properly without losing any button presses
 
+**Fixes**:
+- macOS: Send keyboard key even if the trainer app isn't in foreground
+
 ### 4.7.0 (04-02-2026)
 
 **Features**: